rf(lAnalytics): SuperJson integration; User-specific data removed from side-props
* SuperJson integrated as TRPC transformer to support Date and further types during (de-)serialization
* User-specific data removed from side-props as this does not belong here (side probs are for search engine optimization etc.)
* Support for loading box to indicate long data loading (current seed has already approx 1.7 MB which reduces performance)
