+++ apps/site/pages/_app.tsx
@@ -15,6 +15,7 @@ import { useEffect } from "react";
 import { init } from "@socialgouv/matomo-next";
 import PlausibleProvider from "next-plausible";
 import { LearningAnalyticsSession } from "@self-learning/learning-analytics";
+import superjson from "superjson";
 
 export default withTRPC<AppRouter>({
 	config() {
@@ -29,6 +30,7 @@ export default withTRPC<AppRouter>({
 					url: `${process.env.NEXT_PUBLIC_BASE_PATH ?? ""}/api/trpc`
 				})
 			],
+			transformer: superjson,
 			queryClientConfig: {
 				defaultOptions: {
 					queries: {
